requirements.txt
flask
openai
faiss-cpu
numpy
sentence-transformers
python-dotenv
requests